Revert "move: ContentBox.tsx → customize/ContentBox.tsx"
This reverts commit afc5fd6c40d809db21e717cb312f146b8e7021bf.
