+++ src/components/ContentBox.tsx
